+++ src/test/java/be/robbevanherck/javafraggenescan/dummies/DummyHMMParameters.java
@@ -2,6 +2,8 @@ package be.robbevanherck.javafraggenescan.dummies;
 
 import be.robbevanherck.javafraggenescan.entities.*;
 
+import java.math.BigDecimal;
+
 public class DummyHMMParameters extends HMMParameters {
     /**
      * Create the dummy parameters
@@ -11,30 +13,30 @@ public class DummyHMMParameters extends HMMParameters {
     }
 
     @Override
-    public double getInnerTransitionProbability(HMMInnerTransition transition) {
-        return 0.1;
+    public BigDecimal getInnerTransitionProbability(HMMInnerTransition transition) {
+        return BigDecimal.valueOf(0.1);
     }
 
     @Override
-    public double getInsertInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
-        return 0.2;
+    public BigDecimal getInsertInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
+        return BigDecimal.valueOf(0.2);
     }
 
     @Override
-    public double getMatchInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
-        return 0.3;
+    public BigDecimal getMatchInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
+        return BigDecimal.valueOf(0.3);
     }
 
     @Override
-    public double getOuterTransitionProbability(HMMOuterTransition transition) {
+    public BigDecimal getOuterTransitionProbability(HMMOuterTransition transition) {
         //TODO: make this dependent on the input
-        return 0.4;
+        return BigDecimal.valueOf(0.4);
     }
 
     @Override
-    public double getMatchEmissionProbability(HMMState state, Triple<AminoAcid> aminoAcidEndingInT) {
+    public BigDecimal getMatchEmissionProbability(HMMState state, Triple<AminoAcid> aminoAcidEndingInT) {
         //TODO: make this dependent on the input
-        return 0.5;
+        return BigDecimal.valueOf(0.5);
     }
 
     @Override
